requirements.txt
anthropic==0.83.0
fastapi==0.135.1
fastmcp==3.1.0
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.2
uvicorn==0.41.0
